Refactor: InterviewReminder 관련 객체들을 이메일 도메인으로 패키지 이동
