+++ server/src/main/java/com/ryc/api/v2/email/domain/InterviewReminderSetting.java
@@ -1,4 +1,4 @@
-package com.ryc.api.v2.interview.domain;
+package com.ryc.api.v2.email.domain;
 
 import static com.ryc.api.v2.common.constant.DomainDefaultValues.DEFAULT_INITIAL_ID;
